requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
pydantic-settings==2.1.0
sqlmodel==0.0.27
sqlalchemy==2.0.44
asyncpg==0.29.0
requests==2.31.0
beautifulsoup4==4.12.2
alembic==1.13.1
psycopg2-binary==2.9.9
